requirements.txt
streamlit
numpy
matplotlib
opencv-python
Pillow